refactor(reborn): tighten IronHub composition
